Inventory System Complete  (SEE DESC)
You can now pick up items, they will appear in your inventory, with a name and a description (which we can use for hints).
Will need to adapt this to include a Inspect feature.
1-5 Equips items in that relevant slot.
Q Drops Items.
E Picks items up.

I have a master item we can edit for all assets, so that it should be easy to implement new pick up items moving forward.

Simple UI is opened on the press of I. Will be updated at a later stage, functionality is working though!